requirements.txt
azure-core==1.19.0
azure-ai-textanalytics==5.1.0